refactor: 干掉大杂烩组件 fragments
