+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "@types/classnames": "^2.2.10",
     "@types/jest": "^24.0.0",
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.0",
@@ -18,6 +19,7 @@
     "@types/smooth-scroll": "^16.1.0",
     "@umijs/hooks": "^1.6.2",
     "antd": "^4.0.0",
+    "classnames": "^2.2.6",
     "copy-to-clipboard": "^3.3.1",
     "framer-motion": "^1.8.4",
     "ical-generator": "^1.9.2",
