+++ src/components/GettingStart/index.scss
@@ -1,0 +1,25 @@
+.getting-start-section {
+  min-height: 61.8vw;
+  box-sizing: border-box;
+  text-align: center;
+}
+
+.getting-start {
+  label {
+    color: white !important;
+    font-size: 12pt !important;
+  }
+}
+
+.getting-start-form-mobile {
+  padding-bottom: 1rem;
+}
+
+.getting-start-form-mobile label {
+  display: block;
+  text-align: center;
+}
+
+.getting-start-form-mobile :global(.ant-form-item) {
+  padding-bottom: 1rem;
+}
