+++ src/components/Introduction/assets/introduction-6.png
