+++ src/components/Introduction/assets/types.d.ts
@@ -2,6 +2,6 @@ declare module 'remark-slug';
 declare module 'remark-rehype';
 declare module 'rehype-react';
 declare module 'rehype-react';
-declare module '**/*.css';
-declare module '**/*.jpg';
-declare module '**/*.png';
+declare module '*.css';
+declare module '*.jpg';
+declare module '*.png';
