+++ src/components/Introduction/index.scss
@@ -1,0 +1,17 @@
+.intro-section {
+  min-height: 61.8vw;
+  box-sizing: border-box;
+  text-align: center;
+
+  label,
+  .ant-form-item-label > label {
+    color: white;
+  }
+}
+
+.intro-display-img {
+  max-width: 50%;
+  max-height: 60vh;
+  box-sizing: border-box;
+  text-align: center;
+}
