+++ src/components/ResultPage/index.scss
@@ -1,0 +1,25 @@
+.result-success {
+  flex: 1;
+  background: #56ab2f; /* fallback for old browsers */
+  background: linear-gradient(to top, #5fe053, #56ab2f);
+  color: white;
+  padding-top: 6rem;
+  text-align: center;
+
+  h1 {
+    color: inherit;
+  }
+}
+
+.result-fail {
+  flex: 1;
+  background: #622;
+  background: linear-gradient(to top, #533, #d11);
+  color: white;
+
+  h1 {
+    color: inherit;
+    padding-top: 8rem;
+    text-align: center;
+  }
+}
