--- src/fragments.module.css
@@ -1,67 +1,0 @@
-.Section {
-  min-height: 61.8vw;
-  box-sizing: border-box;
-  text-align: center;
-}
-
-.DisplayImage {
-  max-height: 380px;
-}
-
-.HelpDocContainer {
-  display: flex;
-  flex-direction: row;
-  padding-left: 2rem;
-}
-
-.HelpDocContainer nav {
-  min-width: 25vw;
-}
-
-.HelpDocContainer > * {
-  padding-top: 1.5rem;
-}
-
-.GettingStart label {
-  color: white !important;
-  font-size: 12pt !important;
-}
-
-.FormInMobile {
-  padding-bottom: 1rem;
-}
-
-.FormInMobile label {
-  display: block;
-  text-align: center;
-}
-
-.FormInMobile :global(.ant-form-item) {
-  padding-bottom: 1rem;
-}
-
-.Failure {
-  flex: 1;
-  background: #622;
-  background: linear-gradient(to top, #533, #d11);
-  color: white;
-}
-
-.Failure h1 {
-  color: inherit;
-  padding-top: 8rem;
-  text-align: center;
-}
-
-.Success {
-  flex: 1;
-  background: #56ab2f; /* fallback for old browsers */
-  background: linear-gradient(to top, #5fe053, #56ab2f);
-  color: white;
-  padding-top: 6rem;
-  text-align: center;
-}
-
-.Success h1 {
-  color: inherit;
-}
